refactor(rpc): remove duplicate code
